Added support for what we use from dbt_utils
There are some macros built into the dbt_utils model which results in
queries being made to the data at rest, these tables may not exist yet
if they are upstream of the model. Added an `execute` boolean to track
  if the models are just being compiled or if they are being executed.

If they are just being compiled then those dynamic queries will not run
but instead flag the model as having "Dynamic SQL". When the DAG is
executing any models it encounters with the dyanmic SQL flag will be
recompiled, this time with the execute flag set to true allowing the
dyanmic SQL to be generated.