Adding `isolate-dag` command
This command allows us to easily create a slimmer version of a large DBT
project, which can then be executed by FishTown's DBT which it needing
to parse the whole of your project. i.e. you can get the speed
improvement of DDBT, but the compability of DBT.

This commit also moves the build status updates from Stdout to Stderr